Adding 4 FAQs from Cloud Support
I am proposing 4 new questions identified by a Storage support engineer as ones that they receive frequently. I tried to group them with similar existing questions. These additions are related to a Q2 project to add this service as one for which the Cloud Assistant chatbot provides answers.  Therefore, I also added a #tag (or section ID) for each FAQ because the chatbot needs that metadata.  Thank you.
(I did not include the {: support} tag because I'm not aware of the criteria you use for that one.)